Add consul-based auto-discovery-bootstrap juju.

+++ nubis/bin/consul-autojoin.sh
@@ -5,8 +5,30 @@
 eval `ec2metadata --user-data`
 
 INSTANCE_ID=`ec2metadata --instance-id`
+REGION=`curl -s http://169.254.169.254/latest/dynamic/instance-identity/document | jq '.region' -r`
 
 LOGGER="logger --stderr --priority local7.info --tag nubis-startup"
+
+# For now, we assume CONSUL_DC == REGION
+# XXX: but will need to eventually change to REGION + ENVIRONMENT (i.e. sandbox-us-west-2)
+CONSUL_DC=$REGION
+
+#XXX: For now, default to nubis.allizom.org if not told otherwise, it's a transition path
+#XXX: This is more or less a 1-to-1 mapping from environment name to domain, should be
+#XXX: made discoverable / derivable somehow
+if [ "$NUBIS_DOMAIN" ]; then
+  NUBIS_DOMAIN="nubis.allizom.org"
+fi
+
+# We assume these follow the standard naming scheme...
+CONSUL_UI="http://ui.$REGION.consul.$NUBIS_ENVIRONMENT.$NUBIS_DOMAIN"
+CONSUL_JOIN="$REGION.consul.$NUBIS_ENVIRONMENT.$NUBIS_DOMAIN"
+
+# Auto-discover secret
+SECRET=`curl -s $CONSUL_UI/v1/kv/environments/$NUBIS_ENVIRONMENT/global/consul/secret?raw`
+if [ "$SECRET" ]; then
+  CONSUL_SECRET=$SECRET
+fi
 
 cat <<EOF | tee /etc/consul/zzz-startup.json
 {
@@ -17,6 +39,7 @@ cat <<EOF | tee /etc/consul/zzz-startup.json
 EOF
 
 #XXX: Wish there was a way to self-discover members of our autoscaling group...
+#XXX: For now, this points to the consul bootstrap node (SPOF)
 if [ "$CONSUL_JOIN" ]; then
 cat <<EOF | tee /etc/consul/zzz-join.json
 {
@@ -25,16 +48,7 @@ cat <<EOF | tee /etc/consul/zzz-join.json
 EOF
 fi
 
-if [ "$CONSUL_PUBLIC" -eq "1" ]; then
-PUBLIC_IP=`ec2metadata --public-ipv4`
-
-cat <<EOF | tee /etc/consul/zzz-public.json
-{
-  "advertise_addr": "$PUBLIC_IP"
-}
-EOF
-fi
-
+# This is only used by Consul servers (needs moving to nubis-consul)
 if [ "$CONSUL_BOOTSTRAP_EXPECT" ]; then
 cat <<EOF | tee /etc/consul/zzz-bootstrap.json
 {
@@ -43,19 +57,29 @@ cat <<EOF | tee /etc/consul/zzz-bootstrap.json
 EOF
 fi
 
-if [ "$CONSUL_CERT" ]; then
+# Auto-discover certificate and key
+curl -s -o /etc/consul/consul.pem $CONSUL_UI/v1/kv/environments/$NUBIS_ENVIRONMENT/global/consul/cert?raw
+curl -s -o /etc/consul/consul.key $CONSUL_UI/v1/kv/environments/$NUBIS_ENVIRONMENT/global/consul/key?raw
+
+if [ "$CONSUL_CERT" ] && [ ! -f /etc/consul/consul.pem ]; then
   echo $CONSUL_CERT | tr " " "\n" | perl -pe 's/--BEGIN\n/--BEGIN /g' | perl -pe 's/--END\n/--END /g' > /etc/consul/consul.pem
+fi
+
+if [ -f /etc/consul/consul.pem ]; then
   chown root:consul /etc/consul/consul.pem
   chmod 640 /etc/consul/consul.pem
 fi
 
-if [ "$CONSUL_KEY" ]; then
+if [ "$CONSUL_KEY" ] && [ ! -f /etc/consul/consul.key ]; then
   echo $CONSUL_KEY | tr " " "\n" | perl -pe 's/--(BEGIN|END)\n/--$1 /m' | perl -pe 's/ PRIVATE\n/ PRIVATE /g' > /etc/consul/consul.key
+fi
+
+if [ -f /etc/consul/consul.key ]; then
   chown root:consul /etc/consul/consul.key
   chmod 640 /etc/consul/consul.key
 fi
 
-if [ "$CONSUL_KEY" ] && [ "$CONSUL_CERT" ]; then
+if [ -f /etc/consul/consul.key ] && [ -f /etc/consul/consul.pem ]; then
 cat <<EOF | tee /etc/consul/zzz-tls.json
 {
   "ca_file"         : "/etc/consul/consul.pem",
